create program router at index
